+++ packages/pigeon/example/app/ios/Runner/Messages.g.swift
@@ -403,7 +403,7 @@ class MessageFlutterApi: MessageFlutterApiProtocol {
         let message: String? = nilOrValue(listResponse[1])
         let details: String? = nilOrValue(listResponse[2])
         completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else if listResponse[0] == nil {
+      } else if listResponse[0] == nil || listResponse[0] is NSNull {
         completion(
           .failure(
             PigeonError(