+++ packages/pigeon/platform_tests/test_plugin/darwin/test_plugin/Sources/test_plugin/CoreTests.gen.swift
@@ -5012,7 +5012,7 @@ class FlutterIntegrationCoreApi: FlutterIntegrationCoreApiProtocol {
         let message: String? = nilOrValue(listResponse[1])
         let details: String? = nilOrValue(listResponse[2])
         completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else if listResponse[0] == nil {
+      } else if listResponse[0] == nil || listResponse[0] is NSNull {
         completion(
           .failure(
             PigeonError(
@@ -5072,7 +5072,7 @@ class FlutterIntegrationCoreApi: FlutterIntegrationCoreApiProtocol {
         let message: String? = nilOrValue(listResponse[1])
         let details: String? = nilOrValue(listResponse[2])
         completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else if listResponse[0] == nil {
+      } else if listResponse[0] == nil || listResponse[0] is NSNull {
         completion(
           .failure(
             PigeonError(
@@ -5132,7 +5132,7 @@ class FlutterIntegrationCoreApi: FlutterIntegrationCoreApiProtocol {
         let message: String? = nilOrValue(listResponse[1])
         let details: String? = nilOrValue(listResponse[2])
         completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else if listResponse[0] == nil {
+      } else if listResponse[0] == nil || listResponse[0] is NSNull {
         completion(
           .failure(
             PigeonError(
@@ -5160,7 +5160,7 @@ class FlutterIntegrationCoreApi: FlutterIntegrationCoreApiProtocol {
         let message: String? = nilOrValue(listResponse[1])
         let details: String? = nilOrValue(listResponse[2])
         completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else if listResponse[0] == nil {
+      } else if listResponse[0] == nil || listResponse[0] is NSNull {
         completion(
           .failure(
             PigeonError(
@@ -5188,7 +5188,7 @@ class FlutterIntegrationCoreApi: FlutterIntegrationCoreApiProtocol {
         let message: String? = nilOrValue(listResponse[1])
         let details: String? = nilOrValue(listResponse[2])
         completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else if listResponse[0] == nil {
+      } else if listResponse[0] == nil || listResponse[0] is NSNull {
         completion(
           .failure(
             PigeonError(
@@ -5216,7 +5216,7 @@ class FlutterIntegrationCoreApi: FlutterIntegrationCoreApiProtocol {
         let message: String? = nilOrValue(listResponse[1])
         let details: String? = nilOrValue(listResponse[2])
         completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else if listResponse[0] == nil {
+      } else if listResponse[0] == nil || listResponse[0] is NSNull {
         completion(
           .failure(
             PigeonError(
@@ -5244,7 +5244,7 @@ class FlutterIntegrationCoreApi: FlutterIntegrationCoreApiProtocol {
         let message: String? = nilOrValue(listResponse[1])
         let details: String? = nilOrValue(listResponse[2])
         completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else if listResponse[0] == nil {
+      } else if listResponse[0] == nil || listResponse[0] is NSNull {
         completion(
           .failure(
             PigeonError(
@@ -5275,7 +5275,7 @@ class FlutterIntegrationCoreApi: FlutterIntegrationCoreApiProtocol {
         let message: String? = nilOrValue(listResponse[1])
         let details: String? = nilOrValue(listResponse[2])
         completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else if listResponse[0] == nil {
+      } else if listResponse[0] == nil || listResponse[0] is NSNull {
         completion(
           .failure(
             PigeonError(
@@ -5303,7 +5303,7 @@ class FlutterIntegrationCoreApi: FlutterIntegrationCoreApiProtocol {
         let message: String? = nilOrValue(listResponse[1])
         let details: String? = nilOrValue(listResponse[2])
         completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else if listResponse[0] == nil {
+      } else if listResponse[0] == nil || listResponse[0] is NSNull {
         completion(
           .failure(
             PigeonError(
@@ -5333,7 +5333,7 @@ class FlutterIntegrationCoreApi: FlutterIntegrationCoreApiProtocol {
         let message: String? = nilOrValue(listResponse[1])
         let details: String? = nilOrValue(listResponse[2])
         completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else if listResponse[0] == nil {
+      } else if listResponse[0] == nil || listResponse[0] is NSNull {
         completion(
           .failure(
             PigeonError(
@@ -5364,7 +5364,7 @@ class FlutterIntegrationCoreApi: FlutterIntegrationCoreApiProtocol {
         let message: String? = nilOrValue(listResponse[1])
         let details: String? = nilOrValue(listResponse[2])
         completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else if listResponse[0] == nil {
+      } else if listResponse[0] == nil || listResponse[0] is NSNull {
         completion(
           .failure(
             PigeonError(
@@ -5394,7 +5394,7 @@ class FlutterIntegrationCoreApi: FlutterIntegrationCoreApiProtocol {
         let message: String? = nilOrValue(listResponse[1])
         let details: String? = nilOrValue(listResponse[2])
         completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else if listResponse[0] == nil {
+      } else if listResponse[0] == nil || listResponse[0] is NSNull {
         completion(
           .failure(
             PigeonError(
@@ -5425,7 +5425,7 @@ class FlutterIntegrationCoreApi: FlutterIntegrationCoreApiProtocol {
         let message: String? = nilOrValue(listResponse[1])
         let details: String? = nilOrValue(listResponse[2])
         completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else if listResponse[0] == nil {
+      } else if listResponse[0] == nil || listResponse[0] is NSNull {
         completion(
           .failure(
             PigeonError(
@@ -5456,7 +5456,7 @@ class FlutterIntegrationCoreApi: FlutterIntegrationCoreApiProtocol {
         let message: String? = nilOrValue(listResponse[1])
         let details: String? = nilOrValue(listResponse[2])
         completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else if listResponse[0] == nil {
+      } else if listResponse[0] == nil || listResponse[0] is NSNull {
         completion(
           .failure(
             PigeonError(
@@ -5487,7 +5487,7 @@ class FlutterIntegrationCoreApi: FlutterIntegrationCoreApiProtocol {
         let message: String? = nilOrValue(listResponse[1])
         let details: String? = nilOrValue(listResponse[2])
         completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else if listResponse[0] == nil {
+      } else if listResponse[0] == nil || listResponse[0] is NSNull {
         completion(
           .failure(
             PigeonError(
@@ -5518,7 +5518,7 @@ class FlutterIntegrationCoreApi: FlutterIntegrationCoreApiProtocol {
         let message: String? = nilOrValue(listResponse[1])
         let details: String? = nilOrValue(listResponse[2])
         completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else if listResponse[0] == nil {
+      } else if listResponse[0] == nil || listResponse[0] is NSNull {
         completion(
           .failure(
             PigeonError(
@@ -5549,7 +5549,7 @@ class FlutterIntegrationCoreApi: FlutterIntegrationCoreApiProtocol {
         let message: String? = nilOrValue(listResponse[1])
         let details: String? = nilOrValue(listResponse[2])
         completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else if listResponse[0] == nil {
+      } else if listResponse[0] == nil || listResponse[0] is NSNull {
         completion(
           .failure(
             PigeonError(
@@ -5580,7 +5580,7 @@ class FlutterIntegrationCoreApi: FlutterIntegrationCoreApiProtocol {
         let message: String? = nilOrValue(listResponse[1])
         let details: String? = nilOrValue(listResponse[2])
         completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else if listResponse[0] == nil {
+      } else if listResponse[0] == nil || listResponse[0] is NSNull {
         completion(
           .failure(
             PigeonError(
@@ -5611,7 +5611,7 @@ class FlutterIntegrationCoreApi: FlutterIntegrationCoreApiProtocol {
         let message: String? = nilOrValue(listResponse[1])
         let details: String? = nilOrValue(listResponse[2])
         completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else if listResponse[0] == nil {
+      } else if listResponse[0] == nil || listResponse[0] is NSNull {
         completion(
           .failure(
             PigeonError(
@@ -5642,7 +5642,7 @@ class FlutterIntegrationCoreApi: FlutterIntegrationCoreApiProtocol {
         let message: String? = nilOrValue(listResponse[1])
         let details: String? = nilOrValue(listResponse[2])
         completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else if listResponse[0] == nil {
+      } else if listResponse[0] == nil || listResponse[0] is NSNull {
         completion(
           .failure(
             PigeonError(
@@ -5673,7 +5673,7 @@ class FlutterIntegrationCoreApi: FlutterIntegrationCoreApiProtocol {
         let message: String? = nilOrValue(listResponse[1])
         let details: String? = nilOrValue(listResponse[2])
         completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else if listResponse[0] == nil {
+      } else if listResponse[0] == nil || listResponse[0] is NSNull {
         completion(
           .failure(
             PigeonError(
@@ -5704,7 +5704,7 @@ class FlutterIntegrationCoreApi: FlutterIntegrationCoreApiProtocol {
         let message: String? = nilOrValue(listResponse[1])
         let details: String? = nilOrValue(listResponse[2])
         completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else if listResponse[0] == nil {
+      } else if listResponse[0] == nil || listResponse[0] is NSNull {
         completion(
           .failure(
             PigeonError(
@@ -5732,7 +5732,7 @@ class FlutterIntegrationCoreApi: FlutterIntegrationCoreApiProtocol {
         let message: String? = nilOrValue(listResponse[1])
         let details: String? = nilOrValue(listResponse[2])
         completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else if listResponse[0] == nil {
+      } else if listResponse[0] == nil || listResponse[0] is NSNull {
         completion(
           .failure(
             PigeonError(
@@ -5762,7 +5762,7 @@ class FlutterIntegrationCoreApi: FlutterIntegrationCoreApiProtocol {
         let message: String? = nilOrValue(listResponse[1])
         let details: String? = nilOrValue(listResponse[2])
         completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else if listResponse[0] == nil {
+      } else if listResponse[0] == nil || listResponse[0] is NSNull {
         completion(
           .failure(
             PigeonError(
@@ -6328,7 +6328,7 @@ class FlutterIntegrationCoreApi: FlutterIntegrationCoreApiProtocol {
         let message: String? = nilOrValue(listResponse[1])
         let details: String? = nilOrValue(listResponse[2])
         completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else if listResponse[0] == nil {
+      } else if listResponse[0] == nil || listResponse[0] is NSNull {
         completion(
           .failure(
             PigeonError(
@@ -6461,7 +6461,7 @@ class FlutterSmallApi: FlutterSmallApiProtocol {
         let message: String? = nilOrValue(listResponse[1])
         let details: String? = nilOrValue(listResponse[2])
         completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else if listResponse[0] == nil {
+      } else if listResponse[0] == nil || listResponse[0] is NSNull {
         completion(
           .failure(
             PigeonError(
@@ -6489,7 +6489,7 @@ class FlutterSmallApi: FlutterSmallApiProtocol {
         let message: String? = nilOrValue(listResponse[1])
         let details: String? = nilOrValue(listResponse[2])
         completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else if listResponse[0] == nil {
+      } else if listResponse[0] == nil || listResponse[0] is NSNull {
         completion(
           .failure(
             PigeonError(